# docs/swagger.yaml
definitions:
...
  models.AssetsExpansion:
    properties:
      assets_id:
        description: type:*int     comment:资产id      version:2023-9-20 14:33
        type: integer
      config_category:
        type: string
      group_id:
        type: string
      id:
        description: type:*int     comment:主键        version:2023-9-20 11:23
        type: integer
      name:
        type: string
      name_cn:
        type: string
      property_category:
        type: string
      value:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/n9e/xh/assets-expansion/:
    get:
      consumes:
      - application/json
      description: 根据条件查询资产扩展-西航
      parameters:
      - description: 返回条数
        in: query
        name: limit
        type: integer
      - description: 查询条件
        in: query
        name: query
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AssetsExpansion'
            type: array
      security:
      - ApiKeyAuth: []
      summary: 查询资产扩展-西航
      tags:
      - 资产扩展-西航
    post:
      consumes:
      - application/json
      description: 创建资产扩展-西航
      parameters:
      - description: add assetsExpansion
        in: body
        name: body
        required: true
        schema:
          items:
            $ref: '#/definitions/models.AssetsExpansion'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: 创建资产扩展-西航
      tags:
      - 资产扩展-西航
    put:
      consumes:
      - application/json
      description: 更新资产扩展-西航
      parameters:
      - description: update assetsExpansion
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.AssetsExpansion'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: 更新资产扩展-西航
      tags:
      - 资产扩展-西航
  /api/n9e/xh/assets-expansion/{id}:
    delete:
      consumes:
      - application/json
      description: 根据主键删除资产扩展-西航
      parameters:
      - description: 主键
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: 删除资产扩展-西航
      tags:
      - 资产扩展-西航
    get:
      consumes:
      - application/json
      description: 根据主键获取资产扩展-西航
      parameters:
      - description: 主键
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AssetsExpansion'
      security:
      - ApiKeyAuth: []
      summary: 获取资产扩展-西航
      tags:
      - 资产扩展-西航
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"